update literalincluded snippets
