+++ examples/node/expo-app/AppWrapper.tsx
@@ -1,7 +1,10 @@
-import { Task } from "./app/models/Task";
-
+import LoadingSpinner from "./app/components/LoadingSpinner";
+import LoginUserScreen from "./app/components/LoginUserScreen";
 import App from "./App";
+import {appId} from "./realm.json";
 
+
+const app = new Realm.App({ id: appId });
 
 // :snippet-start: import-task-context
 import TaskContext from "./app/models/Task";
@@ -10,9 +13,6 @@ const { RealmProvider } = TaskContext;
 
 // :snippet-start: wrap-app-within-realm-provider
 function AppWrapper() {
-  // if (!app.currentUser) {
-  //   return <LoginUserScreen />;
-  // }
   return (
     <RealmProvider>
       <App />
@@ -22,3 +22,27 @@ function AppWrapper() {
 // :snippet-end:
 
 export default AppWrapper;
+
+// :snippet-start: dynamically-update-realm-config
+// :replace-start: {
+//   "terms": {
+//     "AppWrapper2": "AppWrapper"
+//   }
+// }
+function AppWrapper2() {
+  if (!app.currentUser) {
+    return <LoginUserScreen />;
+  }
+  const syncConfig = {
+    user: app.currentUser,
+    partitionValue: "ExpoTemplate",
+  };
+
+  return (
+    <RealmProvider sync={syncConfig} fallback={() => <LoadingSpinner />}>
+      <App />
+    </RealmProvider>
+  );
+}
+// :replace-end:
+// :snippet-end:
